Extract DataProductDescriptionField into its own file
Requested by reviewer (Shailesh) on the PR - it was previously a
module-scope component sibling to useDomainCardTemplates. Pure move,
no behavior change: the component still needs its own identity for
per-card useState/useRef/useEffect since renderDataProductCard is a
plain callback invoked once per entity in EntityCardView's loop.
