+++ openmetadata-ui/src/main/resources/ui/src/components/common/atoms/domain/ui/useDomainCardTemplates.tsx
@@ -17,17 +17,16 @@ import {
   Grid,
   Typography,
 } from '@openmetadata/ui-core-components';
-import { ReactNode, useCallback, useEffect, useRef, useState } from 'react';
+import { ReactNode, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { NO_DATA_PLACEHOLDER } from '../../../../../constants/constants';
 import { DataProduct } from '../../../../../generated/entity/domains/dataProduct';
 import { Domain } from '../../../../../generated/entity/domains/domain';
-import { isDescriptionContentEmpty } from '../../../../../utils/BlockEditorPureUtils';
 import { getEntityName } from '../../../../../utils/EntityNameUtils';
 import { getEntityAvatarProps } from '../../../../../utils/IconUtils';
 import { renderBreakableTooltip } from '../../../../../utils/TooltipUtils';
 import { OwnerLabel } from '../../../OwnerLabel/OwnerLabel.component';
-import RichTextEditorPreviewerV1 from '../../../RichTextEditor/RichTextEditorPreviewerV1';
+import { DataProductDescriptionField } from './DataProductDescriptionField';
 import {
   CARD_NAME_CLIP_CLASS,
   CLIPPED_NAME_CLASS,
@@ -36,79 +35,6 @@ import {
   renderDomainOwnersCell,
   renderDomainTypeCell,
 } from './domainFieldRenderers';
-
-/**
- * Description field for the Data Product grid card: a 2-line-clamped
- * markdown-rendered preview, with a "View more" affordance shown only when
- * the text actually overflows 2 lines. Clicking it does nothing on its own —
- * it relies on the card's own onClick (wired in EntityCardView) to navigate
- * into the Data Product, same as clicking anywhere else on the card.
- */
-const DataProductDescriptionField = ({
-  description,
-}: {
-  description?: string;
-}) => {
-  const { t } = useTranslation();
-  const containerRef = useRef<HTMLDivElement>(null);
-  const [isTruncated, setIsTruncated] = useState(false);
-
-  const checkTruncation = useCallback(() => {
-    const container = containerRef.current;
-    if (!container) {
-      return;
-    }
-    // The clamp CSS below targets the nested `.markdown-parser` node that
-    // RichTextEditorPreviewerV1/BlockEditor renders into, not this wrapper -
-    // measure that node (falling back to the wrapper), matching the same
-    // technique FieldCard.tsx already uses for this exact problem.
-    const measureNode =
-      container.querySelector<HTMLElement>('.markdown-parser') ?? container;
-    setIsTruncated(measureNode.scrollHeight > measureNode.clientHeight + 1);
-  }, []);
-
-  useEffect(() => {
-    checkTruncation();
-  }, [description, checkTruncation]);
-
-  useEffect(() => {
-    const container = containerRef.current;
-    if (!container || typeof ResizeObserver === 'undefined') {
-      return;
-    }
-    // RichTextEditorPreviewerV1 renders its BlockEditor lazily (React.lazy +
-    // Suspense), so the real content - and its real height - can land a tick
-    // after this component mounts. The effect above can catch a stale
-    // (pre-load) measurement; this observer re-checks whenever the rendered
-    // height actually changes, which is what BlockEditor's async mount does.
-    const observer = new ResizeObserver(checkTruncation);
-    observer.observe(container);
-
-    return () => observer.disconnect();
-  }, [checkTruncation]);
-
-  if (isDescriptionContentEmpty(description ?? '')) {
-    return <Typography size="text-sm">{NO_DATA_PLACEHOLDER}</Typography>;
-  }
-
-  return (
-    <Box direction="col" gap={1}>
-      <div
-        className="tw:text-sm tw:break-words tw:[&_.markdown-parser]:line-clamp-2"
-        ref={containerRef}>
-        <RichTextEditorPreviewerV1
-          enableSeeMoreVariant={false}
-          markdown={description ?? ''}
-        />
-      </div>
-      {isTruncated && (
-        <Typography className="tw:text-brand-secondary" size="text-xs">
-          {t('label.view-more')}
-        </Typography>
-      )}
-    </Box>
-  );
-};
 
 export const useDomainCardTemplates = () => {
   const { t } = useTranslation();
